Add files via upload
Reverse Linked List in Groups of K

This project demonstrates how to reverse a singly linked list in groups of size k using Python. It includes:

Basic implementation of a singly linked list (Linkedlist and Node classes).

A method to reverse the list in groups of size k.

Sample driver code to test the implementation.

🔁 Problem Statement

Given a linked list, reverse the nodes of the list k at a time and return the modified list.

Only the nodes themselves may be changed.

The function should reverse every k nodes. If the number of nodes is not a multiple of k, the remaining nodes at the end should stay as they are.